+++ tests/utils.py
@@ -46,7 +46,7 @@ def assert_files_equal(path1, path2):
 	try:
 		subprocess.check_output(['diff', '-u', path1, path2], stderr=subprocess.STDOUT)
 	except subprocess.CalledProcessError as e:
-		raise FilesDifferent('\n' + e.output.decode())
+		raise FilesDifferent('\n' + e.output.decode())  # TODO add "from None"
 	except AttributeError:  # Python 2.6 does not have check_output
 		assert subprocess.call(['diff', '-u', path1, path2]) == 0